requirements.txt
nicegui
sqlalchemy
redis
python-dotenv
